requirements/xxx.txt
Fabric==1.4.2
GitPython==0.1.6
Jinja2==2.6
-e git://github.com/mdp-toolkit/mdp-toolkit@a1fae87a932fa9098ec21b4bcaa8c448d7cb5e7e#egg=MDP-dev
MySQL-python==1.2.3
PIL==1.1.7
Pinax==0.9a2
PyOpenGL==3.0.1b2
Pyro==3.14
RunSnakeRun==2.0.2a1
SciTools==0.8
Sphinx==1.1.2
SpikEval==0.6.3
SpikeDB==0.1
SpikePlot==0.1.3
SpikePy==0.2.1
SpikeSort==0.1
SquareMap==1.0.1b3
amqplib==1.0.2
anyjson==0.3.1
argparse==1.2.1
celery==2.5.3
cnetworkmanager==0.21.1
distribute==0.6.24
django-announcements==0.1.2
django-appconf==0.5
django-celery==2.5.5
django-compressor==1.0.1
django-debug-toolbar==0.9.4
django-email-confirmation==0.2
django-extensions==0.8
django-forms-bootstrap==2.0.3.post1
django-mailer==0.2a1
django-notification==0.2
django-openid==0.3a1
django-pagination==1.0.7
django-picklefield==0.1.9
django-sortable==0.2.1
django-sorting==0.1
django-staticfiles==1.2.1
django-taggit==0.9.3
django-taggit-templatetags==0.4.6dev
django-templatetag-sugar==0.1
django-timezones==0.2
django-uni-form==0.9.0
dot2tex==2.8.7
gaphas==0.6.0
gaphor==0.15.0
gunicorn==0.14.3
idios==0.2a1
kombu==2.1.8
mailer==0.7
matplotlib==1.1.1rc
metron==0.1
mysql-connector-python==0.3.2-devel
neo==0.2.0
numexpr==1.4.2
pinax-theme-bootstrap==2.0.3
py==1.3.4
pycrypto==2.6
pygraphviz==1.1
python-dateutil==1.5
python-memcached==1.48
python-openid==2.2.5
pytz==2011n
quantities==0.10.1
scikit-learn==0.10
simplegeneric==0.7
simpleyaml==1.0
six==1.1.0
ssh==1.7.14
stgit==0.15
texttable==0.8.0
winpdb==1.4.8
wsgiref==0.1.2
